+++ src/ParcelManager.Core/Entities/Parcel.cs
@@ -5,8 +5,8 @@
         public string ParcelNumber { get; set; } = default!;
         public string RecipientName { get; set; } = default!;
         public string DestinationCountry { get; set; } = default!;
-        public float Weight { get; set; } = default!;
-        public float Price { get; set; } = default!;
+        public decimal Weight { get; set; } = default!;
+        public decimal Price { get; set; } = default!;
 
         public int BagId { get; set; }
         public BagWithParcels Bag { get; set; } = default!;
